requirements.txt
absl-py==2.3.1
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.4
certifi==2025.8.3
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.3.1
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap4==2024.10
cryptography==42.0.5
daphne==4.1.2
deepface==0.0.96
dj-database-url==3.0.1
Django==5.0.4
django-channels==0.7.0
django-crispy-forms==2.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
filelock==3.20.1
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
gast==0.4.0
gdown==5.2.0
google-auth==2.45.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
joblib==1.5.3
keras==2.12.0
libclang==18.1.1
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.0.8
mtcnn==1.0.0
namex==0.1.0
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==4.25.8
psycopg2-binary==2.9.11
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==24.1.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytube3==9.6.4
pytz==2025.2
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.2.0
rsa==4.9.1
scipy==1.15.3
selenium==4.35.0
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
sqlparse==0.5.0
sympy==1.14.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.14.1
tzdata==2024.1
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.4
whitenoise==6.11.0
wrapt==1.14.2
wsproto==1.2.0
yt-dlp==2024.11.4
zope.interface==6.3
